LayoutTests/imported/w3c/web-platform-tests/largest-contentful-paint/web-font-styled-text-resize-swap-subnode-expected.txt

LCP: Web Font Styled Text Resize

FAIL LCP should be updated if the web font styled text resizes to be larger during the swap period assert_equals: There should be 2 LCP entries. The 1st one corresponds to the system font and the 2nd the web font. expected 2 but got 1
